refactor: Introduce SftpCore for improved service structure
